dev (operators, kafka): moved files out of redundant folder
